requirements.txt
absl-py==0.9.0
aiohttp==3.7.3
anyio==2.0.2
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Babel==2.9.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.5
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.6.0
dominate==2.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
imgaug==0.4.0
imutils==0.5.4
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-server==1.2.2
jupyterlab==2.1.2
jupyterlab-pygments==0.1.2
jupyterlab-server==1.1.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
livelossplot==0.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
missingno==0.4.2
mistune==0.8.4
multidict==5.1.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.6
nest-asyncio==1.5.1
networkx==2.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
ptyprocess==0.6.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.18.1
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.14.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
tornado==6.0.4
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.3.3
yarl==1.6.3
zipp==3.1.0
